[[ Bug 16264 ]] Show the prompt in file open dialogs on El Capitan
If the Mac system version is El Capitan or higher, the prompt will
be used to set the title of the OpenPanel rather than the actual
title. This is because El Capitan's OpenPanel will *only* display
the title as a prompt, it will not set both a title and a message.