Fix redraw issues in combo-box fields
Unlike any other field, the text in a combo-box entryfield is
vertically centred. This means that most of the field code that
calculates dirty rects for redraws uses the wrong y offsets and
ends up clipping incorrectly.

This patch is brute-force and overrides the layer_redrawrect
method for fields to redraw the whole field if it is a combo-box
entry field.

There are still some issues, like cursor positioning being slightly
incorrect, but they are much less jarring and problematic than
having text with clipped descenders!