Use a binary executable for the OSX installer stub
OSX 10.11.4 uses a different method for signing non-MachO files as
code which causes signature verification to fail on earlier versions
of OSX. This commit removes the text files from the Contents/MacOS
directory and uses a MachO executable to launch the AppleScript
that opens a Finder window on the installer DMG root. The script
has been moved to the resource directory.