[[ Bug 17216 ]] Reinstate TIFF processing in Mac clipboard
Some applications on Mac only present TIFF as the image format on
the clipboard (Preview included).

As the engine (as a whole) does not understand TIFF data, this
patch makes TIFF data on the clipboard masquerade as PNG data
allowing it to be used in a loss-less way (relative to the original
TIFF encoding).

The TIFF data will still be listed in the rawClipboardData, but
in the fullClipboardData it will come through as image and png.