[Bug 17322] paletteactions: Ensure submenus show checkmarks.
The paletteactions widget was previously constructing menu control
text for submenus in the following format:

```
View Columns/|revTools_columns
	2/|2
	!c3/|3
	4/|4
```

However, this was preventing submenus from showing checkmarks on
Windows.  The text menu, on the main menu bar, was being constructed
with the tabs *after* the check flags.

This patch modifies the paletteactions widget to format submenus as
follows:

```
View Columns/|revTools_columns
!n	2/|2
!c	3/|3
!n	4/|4
```

This appears to correct the problem.