[[ Bug 17411 ]] Release widget instance on del()
When an object is deleted its 'del()' method is called and then
it is put into the deleted objects pool. This means that an
unused reference to the widget's module can remain until the
pool is drained.

This patch ensures that the MCWidget::del() method releases the
widget instance, meaning that unload extension can work immediately
rather than after the next drain.