[Bug 17385] Prevent repeated OS X app bundle verification
OS X 10.11 requires the `.app` directory to be writable to allow
Gatekeeper to set an xattr on the directory.  This was manifesting
itself by repeated signature verifications and "are you sure you
want to run this app" prompts.  To fix, make only the `Contents`
tree of the LiveCode app bundle read-only.