[[ Bug 17130 ]] Make sure LiveCode doesn't keep open handles to images
This patch ensures that the file handle used to fetch the header of
images is closed after header has been read. This is a trade-off between
parsing the header twice and loading and decompressing image data
when it isn't needed. Given that the latter is substantially more costly
than the former, this is a reasonable balance.