[[ Bug 17472 ]] Ensure drag-drop works from all applications on Mac
When a drag-drop operation starts, the system sends a message with
the draggingSource. If the latter is nil then it means that the
operation is coming from another application and, in which case,
the supplied pasteboard needs to supplant the standard drag
pasteboard. This ensures that apps (such as Word) which create their
own drag-drop pasteboard can still send data to LiveCode.