Avoid O(N^2) slowness when creating dictionary release notes
While generating the dictionary release notes, the release notes
builder was creating an two enormous lists of filenames, then
searching for each line in the first among the lines in the second.
This was in order to make sure that only that new-format release notes
were considered for the dictionary release notes.

It would be possible to speed this up a bit by adding a hashtable, but
we can speed this up _a lot_ by simply ignoring all files that don't
have a `.lcdoc` extension.