[[ SB Inclusions ]] Add helper function to translate inclusions into deploy params
Essentially this just abstracts what was already happening on desktop - take the
settings, and build a list of auxiliary stackfiles and a startup script to
initialise them.

It also resaves all the auxiliary stackfiles as non-script only stacks which
prevents a crash on startup.

Finally, it computes the selected lcb extension dependencies and sets the deploy
'modules' param accordingly.