[Bug 14659] libscript: Initialise module slots with default values.
Initialise module-level variable slots with default values, when
possible.  The information about module-level slot types is only
attached to variable definitions, so it's not possible to do just a
single pass over the slots because there's no way to find out from a
slot index what type is expected to live in it.  Instead, the module
level default values are implemented by initializing slots in two
passes, first to set default values for all variables that have types
associated, and second to set all remaining slots to `kMCNull`.