navbar: Workaround to make color properties appear in docs.
The color properties of the navbar don't really exist at the LCB level
-- they are implemented by MCControl.  However, the navbar LCB source
code contains metadata (so that they show up correctly in the PI) and
documentation (so that an explanation of their effects appear in the
dictionary).

To make the documentation appear, it's necessary to move the
documentation into the module-level doc comment block.