[Bug 17692] graph: Prevent setting empty data
Previously, setting the **graphData** property of a graph widget to an
empty string would appear to succeed, but would make the graph widget
throw an error every time the engine tried to paint it, with
appropriately catastrophic results.

This patch simply ensures that it is impossible to set the
**graphData** to an empty string by throwing an error if you attempt
to do so.