[[ Bug 17468 ]] Fix crash when deleting target object
Previously it was possible for a crash to occur because the target
object can be deleted in a frontScript and the deleted object pool
mechanism meant that it would be destroyed whilst still being on
the C stack (in MCObject::message).

This patch fixes the problem by introducing a root pool, and actions
to suspend and resume an object's deletion. When an object's deletion
is suspended it is promoted to the parent pool, thus making it
impossible for it to be deleted until after the current invocation of
MCObject::message completes. The object is demoted back to its original
pool after the message has fallen through the message path.