[[ ReturnValue ]] Provide new return forms to return values and errors
This patch adds two new forms of the 'return' command:

    return value <value>
    return error <value>

These forms act slightly differently, depending on whether the handler
calling them is a command or a function handler.

If the handler performing the return is a command handler then:

The 'return value' form, sets 'it' in the caller to <value> and
sets 'the result' to empty.

The 'return error' form, sets 'it' in the caller to empty and
sets 'the result' to <value>.

If the handler performingthe return is a function handler then:

The 'return value' form, returns <value> as the return value of
the function call and sets 'the result' to empty.

The 'return error' form, returns empty as the return value of
the function call and sets 'the result' to <value>.

These forms of return are designed to be used by script library
functions to allow them to have the same ability as built-in engine
functions and commands - namely the ability to return a value (in it
for commands, or return value for functions) *or* return a status
(in the result).