[[ CompilerTests ]] Add a compiler test runner
This patch adds a new test runner for running LCB compiler tests.

The compiler tests are run by using 'make compiler-check' in the
tests folder.

The compiler test runner recursively searches for tests in the
tests folder, processing any file with a 'compilertest' extension.

Compiler tests are text files containing a number of test clauses.
Each test clause is named, and consists of a fragment of LCB code
to be parsed to the compiler, along with an expectation and list
of errors and/or warnings to check for.

The compiler test format uses lines beginning with %% to indicate
comments.

A test clause has the following structure:

    %TEST <name>
    ... LCB code ...
    %EXPECT ( PASS | FAIL | XPASS | SKIP )
    ... assertions ...
    %ENDTEST

Here the assertion clauses can be one of the following:

    %SUCCESS
    %ERROR <error-id> AT <pos-id> [ WITH <hint> ]
    %WARNING <warning-id> AT <pos-id> [ WITH <hint> ]

In the body of LCB code for each test clause, %{pos-id} can be used
to indicate a named position - these can be used in error and warning
assertions.

When running a test clause, the test runner will first extract the
LCB code, removing any named positions from it. It will then attempt
to compile the code. The output of the compiler is checked against
the list of expected errors and warnings, or success and each test
will succeed or fail as a result, in line with the expectation.

Additionally, there is a new compiler option '--error-info' which
emits error information in a more structured way. Each error or
warning that is emitted has the following format:

    <file>:<line>:<col>:warning|error:<id>:<hint>

Here, rather than use a human readable message, the compiler emits
the error's id, and optional parameter separately.