Add a checked MCObjectCast function
This function checks that a cast is valid (using MCAssert) when
casting MCObject* to a derived class (e.g. MCButton*). It works
by adding a constant in each derived class indicating the Chunk_Term
used to identify it.

I've only added the constant for MCButton and MCField in this commit;
adding the others is trivial.