[[ Bug 17747 ]] Make sure mouseUp is sent before popup is closed
This patch ensures that a mouseUp which causes a popup to close
is sent *before* the menu is closed. This fixes the problem of
menu closure causing a 'mouseCancel' event in widgets which
occurred previously because closemenu() would cause the grabbed
widget to be reset to nil, indicating that the mouse event
was cancelled even though it was not.