[[ tsNet ]] Implement changes required to support tsNet libURL driver
This patch provides support for the inclusion of
tsNet into the livecode package. All additional
externals (mergExt & tsNet) will now be found
in the same `Tools/Ext` directory.

Support has been implemented for a directory within
`Tools/Ext` to contain livecode files that may insert
themselves into the IDE message path and standalone
build process. This was required for tsNet's script
based libURL driver. These files will have
`revLoadLibrary` and `revUnloadLibrary` sent to
them so they may manage their own lifecycle as
back script, front script, library, behavior or
just loaded stack.

In the course of development a bug was identified
and resolved where android externals were potentially
being filtered from the included files. The
`revSBLibrary` was elevated to an always loaded
back script in the IDE due to the many scripts
depending on it.

There are dependencies between this patch and an
associated patch to the livecode-ide repo.