libfoundation: Clean up signed/unsigned comparisons.
Try to ensure that numeric comparisons in libfoundation are always
performed between numbers of the same signedness and, wherever
possible, the exact same representation.  The most common case is a
loop from 0 to a limit (stored in a variable), in which case the loop
variable should be the same type as the limit variable.