Add MCManagedPThread class.
Add a helper class for managing a `pthread_t` value and associated
"liveness" state.

The `pthread_t` type is opaque, and its structure is
implementation-defined.  The POSIX thread specification doesn't
provide any value for a `pthread_t` that is guaranteed to be an
invalid thread identifier, so it can't be validly initialised to
`NULL` or compared with `NULL`.

Similarly, `pthread_t` values can only be safely compared with
`pthread_equal()`, not with `==`.

The new `MCManagedPThread` class provides an implicit `bool` cast
operator so that it can be used directly in `if` statements, provides
`==` operators for other `MCManagedPThread` values and for `pthread_t`
values for comparison purposes, and provides `Join()` and `Create()`
methods that wrap the corresponding pthread API procedures.