Fix many allocation/deallocation mismatches.
The LiveCode engine uses three different allocators with distinct
semantics.

- `malloc()`/`free()`: these are C library procedures, usually
  accessed via the `MCMemory*` libfoundation functions; they allocate
  memory on the heap and do no C++ construction or destruction of
  their contents.

- `new`/`delete`: these C++ operators allocate/deallocate a single
  instance on the heap, and calls its constructor/destructor.

- `new[]`/`delete[]`: these C++ operators allocate/deallocate an sized
  array of instances on the heap, and call the constructor/destructor
  for each of them.

Mismatch (e.g. allocating with `new[]` and freeing with `free()` is a
resource management error, since it can result in something having a
destructor called when it shouldn't (or vice versa).  Both Coverity
Scan and the newest version of valgrind detect mismatched use.

This patch is a mostly mechanical replacement of release statements to
match the corresponding allocators, with addition of annotation of
what the expected allocator is.  (I added the annotation to help me
detect when multiple codepaths were allocating in conflicting
ways). As part of this, I changed a couple of inconsistent allocation
paths.

A particular error that occurs frequently is using an
`MCAutoPointer<T>` to manage a `T*` that was created using `malloc()`;
in particular, the usual error is to try and put the result of a
libfoundation function that allocates a buffer into an
`MCAutoPointer<T>`.  This won't work, because libfoundation always
allocates memory with `malloc()`, and `MCAutoPointer` always destroys
its contents with `delete`.  Several codepaths are corrected to use an
appropriate `MCAutoCustomPointer` instead.