Use libfoundation allocators for arrays that will be resized.
Using `new[]` and `delete[]` for allocating an array prevents resizing
the array by reallocation.  Resizing by reallocation may cause a
memory address change for the array, which could cause problems if
something holds a pointer to the contents of the array.  Special
handling is therefore required if resizing an array allocated with
`new[]`, often involving an explicit memory copy or move.  When it's
known/obvious that it's safe to `realloc()` a buffer, it's better to
use the libfoundation memory handling functions, which will avoid a
copy most of the time.