[[ Bug 17962 ]] Ensure _mainStacksChanged isn't sent to a deleted stack
When deleting a stack the default stack may be left
in a state where it references the deleted stack
which is an issue which needs further thought and
broader changes in a non-maintenance release. This
patch changes the target for `_mainStacksChanged`
which is always sent just after a stack is deleted
to the home stack. `_mainStacksChanged` is an IDE
only message so this change should have no impact
on users.