libfoundation: Dereference AutoStringRefAs* instances as const.
The buffers that the `AutoStringRefAs*` utility classes expose are not
mutable, and dereferencing them with `operator*` shouldn't mutate the
underlying class.

There's one exception: `MCAutoStringRefAsWString` is quite often used
to provide a `LPWSTR` for passing to Windows API functions.  Many of
these `LPWSTR` arguments are _not_ `const`, and having to `const_cast`
many uses of `MCAutoStringRefAsWString` would remove a lot of the
utility of the class.