libfoundation: Improved MCClamp().
Redefine `MCClamp()` to make its return type the same as the type of
its limit arguments.  This better matches the common usage of
`MCClamp()` whereby the it is used as part of a type conversion
process.

This also enables the use of `MCClamp()` for any type which defines
comparison operators and a copy constructor.