[build] Tweaks to toolsBuilderFilterExternals
A regression was introduced to installer construction somewhere
between commit 38cae350 and commit c9ced4b7.  The regression causes
the Windows executable to be omitted from Indy and Business
installers.

Examination of the git logs suggest that commit d3ed540c is the only
change that should have affected the build process.  This patch:

- checks that setting the current working directory is successful
- sets the current working directory back to what it was initially
- tells the `rm` command to print a list of the files that it deletes