Ensure that field color chunks are fully initialized.
Some C++ compilers don't fully implement the C++11 standard's
requirement that value members of structures should be value
initialised by the default (implicit) constructor.

This means that an `MCColor` is not initialised to `{0,0,0}`, and by
extension, `MCInterfaceNamedColor::color` is not zero-initialised by
its default constructor.

This patch adjusts `PodFieldPropType<MCInterfaceNamedColor>::init()`
to explicitly initialise the color to 0.