Hard wrap lcdoc files
This patch wraps lcdoc files at 72 columns where
appropriate and orders lcdoc elements consistently
in all files. It also fixes some file naming
anomalies where `_` or ` ` was used instead of `-`
for space or the filename did not include the entire
name.

Additionally this patch removes web as a platform due
to the confusion between the legacy browser plugin
and html5 which is listed as an OS.