build: Ensure removed dictionary entries don't break the build
Both the dictionary info gathering and the release note fragment
gathering steps use lists of changed files in order to determine which
files to consider.  Some of these files have changed because they've
been removed.  It isn't an error or even a warning if these files are
missing; it's just an interesting and perhaps notable fact.