[Bug 17721] Correctly identify relevant release notes.
In commit b61890af, the release notes builder was adjusted to only
show the changes since the most recent GM release.  This introduced a
bug.

Each tag under consideration was being matched against `item 1 to 3`
of the version for which the notes are being compiled.  This wasn't
quite right; when making notes for `8.1.0-dp-3`, `item 1 to 3` with
`.` as item delimiter matches the whole of `8.1.0-dp-3`.  The result
was that _all_ relevant release notes were being allocated to the
current release.

This patch ensures that tag relevance is computed using only the
numerical part of the current version string.