[Bug 18042] Fix special character escaping in guide PDF
The HTML template for generating the User Guide PDF overrode the
`marked.js` built-in functions for rendering code sections.  This
meant that `<` and `>` characters in MarkDown code blocks were
appearing literally in the output HTML.  Among other problems, this
gave the User Guide PDF the very surprising title of `"
theDataA[theIndex]["Title"] & "`, thanks to an example in the Data
Grid guide!