[[ Bug 18068 ]] Make htmlText attributes unicode-aware
This patch makes the htmlText importer and exporter correctly work
where attributes require unicode text encoding (using entities).

Additionally, it also fixes the use of non-BMP unicode chars in
all areas the htmlText.

Finally, it generalizes the imageSource char field to multiple
chars. This means that imageSource can now cover multiple characters
which is required for it to work with surrogates and complex graphemes.