[[ DontLoadStacks ]] Remove automatic stack open in development mode
This patch removes automatic loading of stackfiles
provided on the command line from the development
engine. This patch requires a patch to the IDE to
load stackFiles explicitly when the environment
is completely initialised. This patch will allow
the IDE to avoid user stacks presenting while the
splash screen is open and to handle stack files
differently depending on their type. For example,
handle script only stacks by opening the script
editor.