Add generation of initialise_required_weak_link_* functions
These can be used when a library is weakly-linked but is still required
in order for the engine to function.

Currently only implemented on Linux where it prints a message to stderr
then quits. Ideally, it would do something noticeable when run from
a GUI, but it is better than nothing.