com.livecode.iterator: Add "repeat for each _ generated in _"
Add a new syntax for LCB `repeat` loops that allows looping over
values generated by an arbitrary LCB function.  The main current
shortcoming is that it is not possible to specify an initialisation
value for the iterator state.

At some point this might be useful for something other than looping
over the Fibonacci sequence.