[[ Scriptifier ]] Implement scriptifier
This patch adds a scriptifier tool that parses a stackFile and exports
non-trivial scripts as script only behavior stacks. It sets the stackFiles
property ensuring the scripts can be found when the stack is loaded. The
script in the original stack is set to empty. For simplicity any control
with an existing behavior is not modified.