[[ iOS 10 Simulator ]] Update revmobile to support the iOS 10 simulator.
The APIs used by reviphone and reviphoneproxy have moved around in Xcode 8. The
class DTiPhoneSimulatorSystemRoot appears to be no longer used. A new class
SimServiceContext has appeared. We now fetch a SimRuntime from the correct
SimServiceContext for the current version of Xcode and use that to replace the
functionality previously provided by DTiPhoneSimulatorSystemRoot.