[[ LCB ]] Improve tokenisation of numbers and add hex and bin literals This patch improves the tokenisation of numbers in lc-compile and adds support for hexadecimal and binary integer literals.
The compiler will now complain about any invalid suffixes on numeric
literals, meaning that it is not possible to accidentally elide a
numeric literal with an identifier.

Hexadecimal literals are specified using the prefix "0x".

Binary literals are specified using the prefix "0b".

Real literals have been made more flexible and accurate. In particular,
leading zeros are now checked for in all but the fractional part, and
considering a real number to be of the form:
     Int DecPoint Frac Exp
The allowed forms are:
     Int+DecPoint
     Int+Exp
     Int+DecPoint+Exp
     Int+DecPoint+Frac
     Int+DecPoint+Frac+Exp
     DecPoint+Frac
     DecPoint+Frac+Exp
All other combinations will be flagged as a 'malformed real literal'
error.

Note: As the regex for integers now intersects with that for reals
(e.g. 1e1 could be an integer with suffix, or a real) rules for both
DOUBLE and INTEGER literals are now in DOUBLE_LITERAL.t with the DOUBLE
rules first. This ensures DOUBLE_LITERAL matches in preference to
INTEGER_LITERAL.