[[ LCB ]] Add notion of 'bridged' handlers.
The LCB VM currently has some implicit semantics regarding the
calling of foreign handlers which are really only suitable when
the foreign handler has been written to a certain set of rules -
those which the VM expects for its syntax handlers written in C.

In order to explicitly mark a (foreign) handler as conforming to
these rules, this patch adds a new type of handler a 'bridged'
handler. Such a handler is defined using:

    __bridged handler MyHandler() returns nothing binds to "..."

At present bridged handlers only differ from foreign handlers in
that they are implicitly safe.

The notion of bridged handler is intended to be temporary and will
be removed when glue code generators have been written for C
foreign handlers.