lc-compile: Add lazily-constructed array of file lines to File structure
For printing out context of warnings and errors generated by
lc-compile, it's necessary to be able to get access of the text of
each line in the source file.

The position doesn't include a file offset, and scanning the whole
file for each position to find its surrounding context would be
excessively expensive.  Instead, this patch takes the approach of
loading the file's text into a buffer, and then creating an array of
pointers into the buffer that point to the start of each line.

When a file doesn't have any warnings or errors, there's no point in
going to the trouble of setting it up, so the file is scanned lazily
the first time that a particular line's text is needed.