[[ Toolchain ]] Add --outputauxc option to lc-compile
This patch adds an --outputauxc option to lc-compile which can
be used instead of --outputc. It generates modules (encoded in C
arrays) using a global variable g_builtin_aux_modules, rather than
g_builtin_modules. The purpose is to allow additional builtin modules
to be included in a application beyond those included in the engine.