[[ Toolchain ]] Allow 'outputc' modes to process multiple LCB modules
This patch adds the ability for lc-compile to process multiple
LCB modules together to produce a C output file.

This is similar to --outputc mode in bootstrap mode, except that
it doesn't require all modules and their dependencies to be present.
Instead, when resolving a use clause, the compiler will first
search the modules on the command line and *then* search for
interface files in the specified modulepath directories.