Use `new (nothrow)` everywhere
We jump through various hoops in order to get new to be recognised as
having nothrow behaviour but it is fragile and doesn't work when the
engine is hosted inside another process (if that process uses its own
implementation of new, as it is entitled to do).

Instead of dealing with this, just bite the bullet and add the
(std::nothrow) argument to every invocation of new in the engine and
externals.