[Valgrind] Fix buffer overrun in MCU_sttrtor8()
The char buffer passed to `MCU_strtor8()` is not guaranteed to be
nul-terminated, which means that `strnlen` rather than `strlen` must
be used if you want measure the length of the string in the buffer.

Fixing this exposed a bug whereby the length `l` and string pointer
`sptr` locals were not being reset after failing to parse an integer
from the buffer.