Enforce indentation with tabs & alignment with spaces
This patch introduces some infrastructure for ensuring a whitespace
style where code is indented with horizontal tab characters and
aligned with spaces.

It relies on the `uncrustify` tool [1].  This is packaged for
Ubuntu and Fedora, can be installed on MacOS using `homebrew`, and
there are Windows binaries available too.

This patch adds `indent` and `check-indent` targets to the
top-level Makefile, which apply indentation rules and verify them,
and adds a Travis CI rule which fails the build if there are any
indentation mismatches.

Initially, _all_ existing source files are blacklisted from
having indentation rules applied, apart from libfoundation-system
and the Emscripten-specific source code.

The scripts do some acrobatics to accommodate both the ancient
version of `uncrustify` available in Ubuntu and the cutting-edge
version installable with `homebrew`.

[1] http://uncrustify.sourceforge.net/