[[Docs]] constant/return - Example and note format
The formatting of the example and the two End Notes in the Description have been fixed. It's mostly aesthetic to the file itself to make it easier to read and edit in a text editor. They seem to display okay in the API Dictionary as is.
(cherry picked from commit 0333a243d905805e60a3511c5d969c64b1737451)