[Bug] Ensure database connections are released during quit
A previous patch missidentified an errant `break;` as being in the
wrong condition when infact it should not have been there at all.
It was causing only a single connection to be released rather than
all.

This patch ensures that for each connection we iterate all databases
to find the associated database and call their `releaseconnectionptr`