[Bug 18579] Implement accept on ip:port
Previously the accept connections command always bound to INADDR_ANY
which is a security risk if the application only needs to bind to
known interfaces. This patch implements a host string parameter
in the same form as the open socket command of ip:port allowing
it to bind to specific interface ports.