Disable "conversion" and "shorten-64-to-32" warnings for Mac/iOS
These warnings are largely correct, but it would require an
enormous amount of work (probably a general refactor of the engine)
to address them properly.  Adding casts of dubious correctness to
silence the warnings would simply make it harder to track down and
fix the issues when the resources become available.

However, currently the Travis CI build logs are being swamped by
more than 10k lines of warnings, making it impossible to review
testsuite results.  This patch turns off the warnings for the time
being.