libfoundation: Add MCSpan class for wrapping pointer+length pairs.
There are many places in the engine where (pointer, length) pairs are
handled separately.  Some bugs have occurred where they get separated
-- for example, when the pointer gets erroneously passed to a function
that expects to receive a null-terminated array rather than an array
with an extrinsic length.

The `MCSpan` class is inspired by the `span` class recommended by the
C++ Core Guidelines [1] and the `gsl::span` class included in
Microsoft's Guideline Support Library [2].  It is intended to be a
zero-cost abstraction that combines pointer + length into a single
class, and defines safe, checked access and subscripting operations.

It's not quite right yet, because to be fully zero-cost it requires
all of its methods to be declared `constexpr` and not all of the
compilers we use support that yet.

[1] https://github.com/isocpp/CppCoreGuidelines
[2] https://github.com/Microsoft/GSL