json: Import a comprehensive JSON minutiae test suite
This commit vendors the JSONTestSuite set of JSON parsing and
transformation tests into the source tree.  These are originally
from https://github.com/nst/JSONTestSuite, and cover a large
variety of obscure minutiae relating to the various standards for
JSON.

They are not imported as a submodule because we have quite enough
of them already

This commit adds the parsing tests to the test suite and makes
provision for skipping problematic or irrelevant tests, and for
marking tests which fail as broken.  There are four classes of
parsing tests which are currently either skipped or broken:

1. Tests which relate to text decoding are skipped.  These are
   failing because LiveCode does not report text decoding errors
   when performing text decoding via `url "file:..."` or
   `textDecode()`.
2. Very deep nesting torture tests are skipped.  Currently LCB
   requires O(N^2) time to construct a `List` of N entries, which
   means causing `JsonImport()` to be run on a sufficiently-deeply
   nested JSON array would be a super effective DoS attack on an
   application that uses `JsonImport()` on untrusted JSON data.
3. There is a bug in number parsing that means that JSON containing
   a "lonely" number outside an array or object (e.g. "25") will
   not parse successfully.  Bug 18697 has been filed.
4. There is a bug in number parsing whereby some number formats
   that are not permitted in JSON are accepted by LiveCode (e.g.
   "0.e1").