Define kMCSOutputTextEncodingNative outside MCSOutputTextEncoding enum
When `kMCSOutputTextEncodingNative` is defined within the
`MCSOutputTextEncoding`, it is not possible to write a `switch`
statement over all members of the enumeration because
`kMCSOutputTextEncodingNative` aliases with one of the other members.

Making `kMCSOutputTextEncodingNative` into a constant defined outside
the enumeration means that it can be used as a default value for a
text encoding, but allows a `switch` statement over
`MCSOutputTextEncoding` to be validated as complete.

Coverity-Id: 16051