libscript: Prevent null dereference on error when property has no getter
The `MCScriptGetTypeOfPropertyInModule()` function could deference a
null pointer when attempting to determine the get type for the
specified property.  This function was only used in the
`MCScriptThrowInvalidValueForPropertyError()` error dispatch helper,
which in turn was only used from contexts where the VM had already
resolved a specific type for the property and had _just_ failed to
compare a value with it.

This patch removes the `MCScriptGetTypeOfPropertyInModule()` function
entirely, and refactors `MCScriptThrowInvalidValueForPropertyError()`
to take the failed property type as an argument.

By doing so it also corrects another bug: when the type required when
setting a property is different from the type obtained by getting the
property, the error message when failing to set the property would be
misleading.

Coverity-Id: 137924