libscript: Don't mix enum types.
The on-disk value saved in the compiled LCB bytecode for the mode of a
handler parameter (`in`/`out`/`inout`) is described by the
`MCScriptHandlerTypeParameterMode` enumeration.  However, when
constructing an `MCHandlerTypeInfo` value for use at runtime, these
values need to be translated to the `MCHandlerTypeFieldMode`
enumeration.  This patch adds an explicit `inline` conversion method
to the `MCScriptHandlerTypeParameter` structure.

Coverity-Id: 16984