Make MCObjectHandle slightly less strict
Get() now no longer asserts that the handle is valid; it instead checks
that it is either unbound or valid. This change has been made to support
a common use pattern in the engine. It should not affect safety as
dereferencing an unbound pointer will cause a segfault.

N.B this pattern is passing nil to methods to indicate no object

Conflicts:
	engine/src/object.h