[Bug 18707] json: Limit maximum nesting depth
As structure nesting depth increases, LCB's current copy-always
List mutation semantics make the cost of starting a Kth level of
structure nesting proportional to K^2.  This starts to really blow
up for very deep nesting and can be used to perform a rather
effective DoS with a relatively small crafted JSON file.

This patch hardcodes a limit of 500 levels of nesting recursion
into the JSON library, which is well before the point where
the cost of nesting starts getting silly.