[Bug 18697] Fix parsing of "lonely number" JSON files
Because JSON numbers are not self-delimiting, it's not possible to
figure out where they end without reading one char further.  This
was causing problems where the char finishes at the end of input.

This patch adds a variable to track the offset into the input
string, and stops the scanner from trying to read off the end of
the input when in a number literal.