com.livecode.assert: Add "expect that _ because _" syntax
This patch adds some new syntax to LiveCode Builder for ergonomically
checking handler preconditions that cannot currently expressed in the
LCB typesystem (e.g. numerical ranges and array structure).  There are
two new statement forms:

    expect that <Condition>
    expect that <Condition> because <Reason>

The `that` token is optional, but you can include it if it makes the
assertion more readable.

My expectation is that this new syntax will be used in the initial
lines of a handler.  For example:

    handler FormatNumberList(in pNumbers as List) returns String
        expect IsListOfNumbers(pNumbers) \
	      because "Input list is expected to contain only numbers"
	/* ... */
    end handler

My use of "expect" rather than "assert" here is an intentional design
decision.  "Assert" is used in English prose in a slightly different
way than typically used in a programming context, whereas "expect" in
this context does exactly what might be expected from the use in
English: it checks the state which the handler expects its argument
(and the larger system) to be in before it's run.

I considered "expects".  I found that it works find as long as the
statements are always at the start of the handler, but an "expects"
statement placed elsewhere looked a bit odd.